requirements.txt
git+https://github.com/pythonforfacebook/facebook-sdk.git
pywebview
itsdangerous
flask